MAINT: Rename Twitter parser to IoC Extractor
Twitter collector is removed (does not work anymore)
but the Twitter parser is still usable, so rename it with a more generic
name
